+++ services/blockstorage/adapter/persistence.go
@@ -20,7 +20,7 @@ type CursorFunc func(first primitives.BlockHeight, page []*protocol.BlockPairCon
 type BlockPersistence interface {
 	WriteNextBlock(blockPair *protocol.BlockPairContainer) (bool, primitives.BlockHeight, error)
 
-	ScanBlocks(from primitives.BlockHeight, pageSize uint8, f CursorFunc) error
+	ScanBlocks(from primitives.BlockHeight, pageSize uint64, f CursorFunc) error
 
 	GetLastBlockHeight() (primitives.BlockHeight, error)
 	GetLastBlock() (*protocol.BlockPairContainer, error)
